+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/attachments/AttachmentsPicker.kt
@@ -182,7 +182,7 @@ public fun AttachmentsPicker(
                     }
 
                     LaunchedEffect(storagePermissionState.hasPermission) {
-                        if (storagePermissionState.permissionRequested && storagePermissionState.hasPermission) {
+                        if (storagePermissionState.hasPermission) {
                             attachmentsPickerViewModel.loadData()
                         }
                     }