+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/header/MessageListHeader.kt
@@ -85,7 +85,7 @@ public fun MessageListHeader(
     },
     trailingContent: @Composable RowScope.() -> Unit = {
         ChannelAvatar(
-            modifier = Modifier.size(36.dp),
+            modifier = Modifier.size(40.dp),
             channel = channel,
             currentUser = currentUser,
             contentDescription = channel.name,